// SafariZoneMKIII/src/app/interfaces/pokedex-entry.ts
export interface PokedexEntry {
   dexNum: number
   pokemon: string
   hp: number
   atk: number
   def: number
   satk: number
   sdef: number
   spd: number
   type: string
   ability1: string
   ability2: string
   EVWorth: string
   gender: string
   evolvesFrom: string
   evolvesBy: string
   evolvesTo: string
   discovered: string
   description: string
}